+++ src/portfolio_forecast/utils/__init__.py
@@ -1,4 +1,5 @@
+from .bar_times import close_times
 from .periods import PERIODS_PER_YEAR
 from .trading_dates import next_trading_dates
 
-__all__ = ["PERIODS_PER_YEAR", "next_trading_dates"]
+__all__ = ["PERIODS_PER_YEAR", "close_times", "next_trading_dates"]